requirements.txt
pyyaml
argparse
numpy
h5py
torch
torchvision
sqlalchemy
psycopg2-binary